Restructure snappy_cpp headers to ROS include/snappy_cpp layout and move THRUSTER_MAP to docs/
