+++ src/snappy_cpp/include/snappy_cpp/serial.h
